+++ sources/plugins/TeleInfo/Keywords/ForecastTomorrow.h
@@ -26,6 +26,12 @@ public:
    void set(const std::string& Value);
 
    //--------------------------------------------------------------
+   /// \brief	                      Set the value of the keyword
+   /// \param[in] Value                The new value
+   //--------------------------------------------------------------
+   void set(const teleInfo::specificHistorizers::EColor& Value);
+
+   //--------------------------------------------------------------
    /// \brief	                      Get if the value have changed
    /// \return                       true, if it's a new value
    //--------------------------------------------------------------